Update MarkQuestionSubpage, ScriptViewSubpage
