+++ frontend/src/AuthenticatedApp/paperScripts/ScriptViewSubpage/index.tsx
@@ -17,60 +17,12 @@ import {
 import ArrowBackIcon from "@material-ui/icons/ArrowBack";
 import ArrowLeftIcon from "@material-ui/icons/ArrowBackIos";
 import ArrowRightIcon from "@material-ui/icons/ArrowForwardIos";
-import { createStyles, makeStyles, Theme } from "@material-ui/core/styles";
-import { lightBlue } from "@material-ui/core/colors";
 
 import { CanvasWithToolbar } from "../../../components/Canvas";
 import LoadingSpinner from "../../../components/LoadingSpinner";
 import ReversedChip from "../../../components/ReversedChip";
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    container: {
-      minHeight: "100vh",
-      minWidth: "100vw",
-      display: "flex",
-      flexDirection: "column",
-      touchAction: "none"
-    },
-    grow: {
-      display: "flex",
-      flexGrow: 1
-    },
-    backButton: {
-      marginRight: theme.spacing(2)
-    },
-    prevPageButton: {
-      position: "absolute",
-      top: 0,
-      bottom: 0,
-      left: 0,
-      margin: "auto"
-    },
-    nextPageButton: {
-      position: "absolute",
-      top: 0,
-      bottom: 0,
-      right: 0,
-      margin: "auto"
-    },
-    pageLabel: {
-      position: "absolute",
-      bottom: theme.spacing(10),
-      left: 0,
-      right: 0,
-      margin: "0 auto"
-    },
-    questionBar: {
-      backgroundColor: lightBlue[50],
-      top: "auto",
-      bottom: 0
-    },
-    questionBarItem: {
-      marginRight: theme.spacing(1)
-    }
-  })
-);
+import useStyles from "./styles";
 
 type Props = RouteComponentProps;
 
@@ -147,21 +99,16 @@ const ScriptView: React.FC<Props> = ({ match: { params } }) => {
 
   const {
     matriculationNumber,
-    rootQuestion,
-    descendantQuestions,
+    rootQuestionTemplate,
+    questions,
     pages
   } = scriptViewData;
 
   const getCurrentPageQuestions = () => {
     const currentPage = pages.find(page => page.pageNo === pageNo)!;
-    const currentPageQuestions =
-      descendantQuestions === undefined ||
-      descendantQuestions === null ||
-      descendantQuestions.length === 0
-        ? [rootQuestion]
-        : descendantQuestions.filter(question =>
-            currentPage.questionIds.includes(question.id)
-          );
+    const currentPageQuestions = questions.filter(question =>
+      currentPage.questionIds.includes(question.id)
+    );
     return currentPageQuestions;
   };
